+++ src/components/basics/DropdownMenuItem/dropdown-menu-item.stories.tsx
@@ -1,9 +1,10 @@
 import type { ComponentMeta, ComponentStory } from "@storybook/react";
-import icons from "../Icon/lib";
+import icons from "@/components/basics/Icon/lib";
 import DropdownMenuItem from "./dropdown-menu-item.component";
+import { DropdownMenuItemSizes } from "./dropdown-menu-item.types";
 
 export default {
-  title: "Components/DropdownMenu/DropdownMenuItem",
+  title: "Components/DropdownMenuItem",
   component: DropdownMenuItem,
   argTypes: {
     leftIconName: {
@@ -14,24 +15,28 @@ export default {
       options: Object.keys(icons),
       control: { type: "select" },
     },
+    size: {
+      options: DropdownMenuItemSizes,
+      control: { type: "select" },
+    },
   },
 } as ComponentMeta<typeof DropdownMenuItem>;
 
 const Template: ComponentStory<typeof DropdownMenuItem> = (args) => <DropdownMenuItem {...args} />;
 
-export const Default = Template.bind({});
-Default.args = {
+const DefaultArgs = {
   children: "Item de Menu",
+  size: DropdownMenuItemSizes.Medium,
 };
+
+export const Default = Template.bind({});
+Default.args = DefaultArgs;
 
 export const LeftIcon = Template.bind({});
-LeftIcon.args = {
-  children: "Item de Menu",
-  leftIconName: "me-salva-mini",
-};
+LeftIcon.args = { ...DefaultArgs, leftIconName: "me-salva-mini" };
 
 export const RightIcon = Template.bind({});
-RightIcon.args = {
-  children: "Item de Menu",
-  rightIconName: "me-salva-mini",
-};
+RightIcon.args = { ...DefaultArgs, rightIconName: "me-salva-mini" };
+
+export const Small = Template.bind({});
+Small.args = { ...DefaultArgs, leftIconName: "trash", size: DropdownMenuItemSizes.Small };