+++ src/components/basics/Icon/icon.component.tsx
@@ -18,10 +18,12 @@ const Icon = ({ name, size = IconSizes.Medium, className, ...props }: IconProps)
 
 const getSize = (size: string) => {
   switch (size) {
+    case IconSizes.XLarge:
+      return 40;
     case IconSizes.Large:
       return 32;
     case IconSizes.Small:
-      return 16;
+      return 18;
     case IconSizes.Medium:
     default:
       return 24;