+++ src/components/basics/SubjectLabel/subject-label.types.ts
@@ -1,5 +1,5 @@
 import { LabelProps, LabelSubjectVariants } from "@/components/basics/Label";
 
 export interface SubjectLabelProps extends Omit<LabelProps, "theme" | "variant"> {
-  variant: LabelSubjectVariants;
+  variant?: LabelSubjectVariants;
 }