+++ src/components/forms/Slider/slider.types.ts
@@ -1,7 +1,10 @@
-export interface SliderProps extends React.HTMLProps<HTMLInputElement> {
+import { HTMLProps } from "react";
+
+export interface SliderProps extends Pick<HTMLProps<HTMLDivElement>, "className" | "style"> {
   max?: number;
   min?: number;
   formatter?: (value: number) => string;
   defaultValue?: number;
   step?: number;
+  onSelect?: (value: number) => void;
 }