+++ CHANGELOG.md
@@ -5,6 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha.18](https://github.com/mesalva/mars/compare/v1.0.0-alpha.17...v1.0.0-alpha.18) (2022-01-05)
+
+
+### Bug Fixes
+
+* components useds in checkout ([#89](https://github.com/mesalva/mars/issues/89)) ([f30b8c0](https://github.com/mesalva/mars/commit/f30b8c059129e92bd2f9ace93ebf431d7d357c85))
+
+## [1.0.0-alpha.17](https://github.com/mesalva/mars/compare/v1.0.0-alpha.16...v1.0.0-alpha.17) (2022-01-04)
+
+
+### Bug Fixes
+
+* credit card component ([#88](https://github.com/mesalva/mars/issues/88)) ([0b40aec](https://github.com/mesalva/mars/commit/0b40aecc248a95a6099e9cb2d8db073ce61dd6a9))
+
+## [1.0.0-alpha.16](https://github.com/mesalva/mars/compare/v1.0.0-alpha.15...v1.0.0-alpha.16) (2022-01-03)
+
+
+### Bug Fixes
+
+* try trigger release ([1249830](https://github.com/mesalva/mars/commit/1249830de3d90ca7b8e4c3f5ff3a307238ea8dbe))
+
 ## [1.0.0-alpha.15](https://github.com/mesalva/mars/compare/v1.0.0-alpha.14...v1.0.0-alpha.15) (2021-12-31)